updated gitignore to exlude src/test, changed mValue in AnnotationLayer from int to float

+++ src/org/creativecoding/audio/analysis/svl/AnnotationLayer.java
@@ -91,7 +91,7 @@ public class AnnotationLayer
 		public long mFrame;
 		
 		@XmlAttribute(name="value")
-		public int mValue;
+		public float mValue;
 		
 		@XmlAttribute(name="duration")
 		public long mDuration;
